requirements.txt
flask>=2.3
pandas>=2.0
numpy>=1.24
plotly>=5.15
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
statsmodels>=0.14
requests>=2.31
beautifulsoup4>=4.12
Pillow>=10.0
pytest>=7.4
